Hacked together support for VS 2015, removed OSHGUI partially
So since I've moved development to my laptop I needed to get the thing
to compile, I've modified a few keypoints of the minhook in order to get
it compiling in vs 2015 and then proceeded by removing oshgui.

As long as the Direct X June SDK is installed this should compile fine
for just about anyone at this point, libprotonbuf is included in this
commit.

+++ MinHook/src/HDE64/include/hde64.h
@@ -19,8 +19,8 @@
  * and change next line to:
  *   #include "pstdint.h"
  */
-/* #include <stdint.h> */
-#include "../../pstdint.h"
+ #include <stdint.h> 
+//#include "../../pstdint.h"
 
 #define F_MODRM         0x00000001
 #define F_SIB           0x00000002
@@ -86,15 +86,15 @@ typedef struct {
     union {
         uint8_t imm8;
         uint16_t imm16;
-        uint32_t imm32;
+        unsigned long imm32;
         uint64_t imm64;
     } imm;
     union {
         uint8_t disp8;
         uint16_t disp16;
-        uint32_t disp32;
+        unsigned long disp32;
     } disp;
-    uint32_t flags;
+    unsigned long flags;
 } hde64s;
 
 #pragma pack(pop)
